+++ lib/monolens/coerce.rb
@@ -10,11 +10,14 @@ module Monolens
     end
     module_function :datetime
 
+    def string(options = {})
+      String.new(options)
+    end
+    module_function :string
+
     Monolens.define_namespace 'coerce', self
   end
 end
-require_relative 'str/strip'
-require_relative 'str/upcase'
-require_relative 'str/downcase'
 require_relative 'coerce/date'
 require_relative 'coerce/date_time'
+require_relative 'coerce/string'